Move to project setup, iron out small bugs. Replace old src dir
