+++ Project.toml
@@ -1,18 +1,52 @@
+name = "mGV"
+uuid = "b1af0ff3-f930-4051-bc19-f789759b7f7a"
+
 [deps]
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
+Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 Atomix = "a9b6321e-bd34-4604-b9c9-b65b8de01458"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 Configurations = "5218b696-f38b-4ac9-8b61-a12ec717816d"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 JuliaC = "acedd4c2-ced6-4a15-accc-2607eb759ba2"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 PackageCompiler = "9b87118b-4619-50d2-8e1e-99f35a4d4d9d"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 Zarr = "0a941bbe-ad1d-11e8-39d9-ab76183a1d99"
+
+[compat]
+AMDGPU = "2.5.1"
+Accessors = "0.1.45"
+Adapt = "4.6.1"
+Atomix = "1.1.3"
+CUDA = "6.2.0"
+Configurations = "0.17.6"
+Dates = "1.11.0"
+JuliaC = "0.3.6"
+KernelAbstractions = "0.9.41"
+LinearAlgebra = "1.12.0"
+MacroTools = "0.5.16"
+NCDatasets = "0.14.15"
+NetCDF = "0.12.2"
+OffsetArrays = "1.17.0"
+PackageCompiler = "2.4.0"
+Parameters = "0.13.1"
+Printf = "1.11.0"
+ProgressMeter = "1.11.0"
+StaticArrays = "1.9.18"
+Statistics = "1.11.1"
+TOML = "1.0.3"
+TimerOutputs = "0.5.29"
+Zarr = "0.10.0"
