--- src/constants/sim_constants.jl
@@ -1,16 +1,0 @@
-const K_L = ft(0.1)
-const VON_KARMAN = ft(0.4)
-const Z2 = ft(10.0)
-const RI_CR = ft(0.2)
-const EMISSIVITY = ft(0.97)
-
-# Thermal conductivity constants
-const KI = ft(2.2)   # Thermal conductivity of ice (W/mK)
-const KW = ft(0.57)  # Thermal conductivity of water (W/mK)
-const KDRY_ORG = ft(0.05)  # Dry thermal conductivity of organic fraction (W/mK)
-const KS_ORG = ft(0.25)  # Thermal conductivity of organic solid (W/mK)
-
-# Ground composition constants
-const ORGANIC_FRAC = ft(0.0)
-const BULK_DENS_ORG = ft(0.0)
-const SOIL_DENS_ORG = ft(0.0)
