+++ src/energy_balance.jl
