--- src/physics/evapotranspiration.jl
@@ -1,586 +1,0 @@
-function aerodynamic_kernel(z0, d0, tsurf, tair, wind, Z2, Kt, gt, Tf, Ric, z_floor, d_floor, w_floor, L2_min, ra_min, ra_max)
-    # 1. Roughness & Effective Height
-    rough = max(z0, z_floor)
-    d_eff = max(Z2 - d0, d_floor)
-
-    # 2. Log-law terms
-    ratio = clamp(d_eff / rough, ft(1.0e-6), ft(1.0e6))
-    L     = log(ratio)
-    L2    = max(L^2, L2_min)
-    a_sq  = (Kt^2) / L2
-    ccoef = ft(49.82) * a_sq * sqrt(ratio)
-
-    # 3. Stability (Richardson Number)
-    w_spd = max(wind, w_floor)
-    Tmean = max(((tair + Tf) + (tsurf + Tf)) * ft(0.5), ft(100.0))
-    
-    Ri_B  = gt * (tair - tsurf) * d_eff / (Tmean * w_spd^2)
-    Ri_B  = clamp(Ri_B, -ft(0.5), Ric)
-
-    # 4. Friction Factor (Fw)
-    Fw_neg = ft(1.0) - (ft(9.4) * Ri_B) / (ft(1.0) + ccoef * sqrt(abs(Ri_B)))
-    Fw_pos = ft(1.0) / (ft(1.0) + ft(4.7) * Ri_B)^2
-    Fw     = ifelse(Ri_B < ft(0.0), Fw_neg, Fw_pos)
-    Fw     = clamp(Fw, ft(1.0e-3), ft(10.0))
-
-    # 5. Final Resistance
-    C_H = max(ft(1.0) * a_sq * Fw, ft(1.0e-6))
-    ra_val = ft(1.0) / (C_H * w_spd)
-    
-    return clamp(ra_val, ra_min, ra_max)
-end
-
-
-function compute_aerodynamic_resistance!(
-    ra, 
-    Z2, d0_gpu, z0_gpu, z0soil_gpu, tsurf, tair_gpu, wind_gpu, cv_gpu
-)
-    # --- 1. Hard Validation (Debugging/Safety), TODO: remove? ---
-    @assert eltype(ra)       == FloatType "Output 'ra' must be FloatType"
-    @assert eltype(tsurf)    == FloatType "Input 'tsurf' must be FloatType"
-    @assert VON_KARMAN       isa FloatType "Constant 'VON_KARMAN' in SimConstants must be FloatType"
-    @assert G                isa FloatType "Constant 'G' in SimConstants must be FloatType"
-
-    # --- 2. Local Constants ---
-    z_floor = ft(1e-3)
-    d_floor = ft(1e-2)
-    w_floor = ft(0.1)
-    ra_min  = ft(1.0)
-    ra_max  = ft(1e5)
-    
-    # Pre-calculate log expression
-    L2_min  = ft(log(1.01)^2)
-    
-    # Cast scalar Z2 once if it isn't already
-    z2T = ft(Z2)
-
-    # --- 3. Grid Dimensions ---
-    N_all   = size(ra, 4)
-    veg_dim = max(N_all - 1, 0)
-
-    # ========================================================================
-    # 1. SOIL TILES (Last Index)
-    # ========================================================================
-    # We pass the global constants (VON_KARMAN, G, T_FREEZE, RI_CR) directly.
-    @views @. ra[:, :, :, N_all:N_all] = aerodynamic_kernel(
-        z0soil_gpu,                
-        d0_gpu[:,:,:,N_all:N_all], 
-        tsurf,                     
-        tair_gpu,                  
-        wind_gpu,                  
-        z2T, VON_KARMAN, G, T_FREEZE, RI_CR, 
-        z_floor, d_floor, w_floor, L2_min, ra_min, ra_max
-    )
-
-    # ========================================================================
-    # 2. VEGETATION TILES (Indices 1:veg_dim)
-    # ========================================================================
-    if veg_dim > 0
-        @views @. ra[:, :, :, 1:veg_dim] = aerodynamic_kernel(
-            z0_gpu[:,:,:,1:veg_dim],   
-            d0_gpu[:,:,:,1:veg_dim],   
-            tsurf,
-            tair_gpu,
-            wind_gpu,
-            z2T, VON_KARMAN, G, T_FREEZE, RI_CR, 
-            z_floor, d_floor, w_floor, L2_min, ra_min, ra_max
-        )
-    end
-
-    return nothing
-end
-
-
-function calculate_net_radiation!(net_rad, swdown_gpu, lwdown_gpu, albedo_gpu, tsurf,
-                                  snow_coverage_gpu=nothing, snow_albedo_gpu=nothing, snow_surf_temp_gpu=nothing)
-
-    if snow_coverage_gpu === nothing
-        @. net_rad = (ft(1.0) - albedo_gpu) * swdown_gpu + lwdown_gpu - EMISSIVITY * SIGMA * (tsurf + ft(273.15))^4
-    else
-        eff_alb(alb, sc, s_alb) = (isnan(sc) || sc <= ft(0.0)) ? alb : (sc * s_alb + (ft(1.0) - sc) * alb)
-        eff_t(ts, sc, s_ts) = (isnan(sc) || sc <= ft(0.0)) ? ts : (sc * s_ts + (ft(1.0) - sc) * ts)
-        
-        @. net_rad = (ft(1.0) - eff_alb(albedo_gpu, snow_coverage_gpu, snow_albedo_gpu)) * swdown_gpu + lwdown_gpu - EMISSIVITY * SIGMA * (eff_t(tsurf, snow_coverage_gpu, snow_surf_temp_gpu) + ft(273.15))^4
-    end
-    
-    return nothing
-end
-
-function calculate_potential_evaporation!(
-    pe,
-    tair_gpu, tair_grid, psurf_gpu, vp_gpu, elev_gpu,
-    net_radiation, aerodynamic_resistance, rarc_gpu, rmin_gpu, LAI_gpu
-)
-    # 1. Setup Type
-    T = eltype(pe)
-    
-    # --- Local Coefficients ---
-    G_COEFF = ft(1628.6)    
-    AIR_C   = ft(0.003486)  
-    EPS     = ft(1.0e-6)      
-
-    # 2. Pre-calculate 2D Meteorological Terms
-    slope       = @. calculate_svp_slope(tair_grid)           # grid-level T
-    latent_heat = @. calculate_latent_heat(tair_grid + T_FREEZE)  # grid-level T
-    gamma_      = @. G_COEFF * (P_STD * exp(-elev_gpu / calculate_scale_height(tair_grid, elev_gpu))) / latent_heat
-    vpd_pa      = @. calculate_vpd(tair_grid, vp_gpu)
-    air_dens_term = @. ((AIR_C * psurf_gpu * PA_PER_KPA) / (T_FREEZE + tair_grid)) * C_P_AIR * calculate_vpd(tair_grid, vp_gpu) * DAY_SEC
-
-
-    veg_indices = 1:(nveg - 1)
-    
-    # Vegetation tiles: PE at minimum canopy resistance (gsm_inv=1)
-    @views @. pe[:, :, :, veg_indices] = max(
-        (slope * (net_radiation[:, :, :, veg_indices] * DAY_SEC) + 
-         (air_dens_term / aerodynamic_resistance[:, :, :, veg_indices])) / 
-        (latent_heat * (slope + gamma_ * (ft(1.0) + 
-         ((rmin_gpu[:, :, :, veg_indices] / max(LAI_gpu[:, :, :, veg_indices], EPS)) + 
-          rarc_gpu[:, :, :, veg_indices]) / aerodynamic_resistance[:, :, :, veg_indices]))),
-        ft(0.0)
-    )
-
-    # Bare Soil Tile (rc=0, compute_pot_evap bare soil PE)
-    @views @. pe[:, :, :, nveg] = max(
-        (slope * (net_radiation[:, :, :, nveg] * DAY_SEC) + (air_dens_term / aerodynamic_resistance[:, :, :, nveg])) / 
-        (latent_heat * (slope + gamma_ * (ft(1.0) + rarc_gpu[:, :, :, nveg] / aerodynamic_resistance[:, :, :, nveg]))),
-        ft(0.0)
-    )
-
-    return nothing
-end
-
-
-function calculate_max_water_storage!(max_water_storage, LAI_gpu, coverage_gpu)
-
-    @. max_water_storage = ifelse(coverage_gpu > ft(1.0e-5), (K_L * LAI_gpu) / coverage_gpu, ft(0.0))
-
-    @. max_water_storage = ifelse(
-        isnan(max_water_storage) | (abs(max_water_storage) > fillvalue_threshold), 
-        ft(0.0), 
-        max_water_storage
-    )
-
-    return nothing
-end
-
-
-@inline function canopy_evap_physics_kernel(
-    ws, max_ws, pot_evap, ra, rarc, 
-    prec, lai, tair, elev, rmin
-)
-
-    # --- 1. Physics Calculations ---
-    # Using your existing project functions
-    slope = calculate_svp_slope(tair)
-    latent_heat = calculate_latent_heat(tair + T_FREEZE)
-
-    scale_height = calculate_scale_height(tair, elev)
-    surface_pressure = ft(101325.0) * exp(-elev / scale_height)
-    gamma_val = ft(1628.6) * surface_pressure / latent_heat
-
-    # --- 2. Resistances ---
-    rc = rmin / max(lai, ft(1e-6))
-    inv_ra = ft(1.0) / max(ra, ft(1e-6))
-
-    # --- 3. Denominators (Penman-Monteith) ---
-    den_w = slope + gamma_val * (ft(1.0) + rarc * inv_ra)
-    den_rc = den_w + (gamma_val * rc * inv_ra)
-
-    E_p_wet = pot_evap * (den_rc / max(den_w, ft(1e-6)))
-
-    # --- 4. VIC Evaporative Scaling (Branchless) ---
-    # The original VIC model uses conditional logic to evaluate if the canopy is fully wet or dry.
-    # We replace that with a continuous mathematical formulation:
-    # 
-    #   Wratio: Evaluates the current canopy water storage as a fraction of its maximum capacity.
-    #           It is safely clamped between 0.0 (completely dry) and 1.0 (fully wet).
-    #
-    #   ra_ratio: Computes the aerodynamic resistance modifier.
-    # 
-    #   canopy_evap_star: Calculates the potential canopy evaporation scaled non-linearly 
-    #                     (to the 2/3rds power) based on how much water is currently present.
-    Wratio = clamp(ws / max(max_ws, ft(1e-6)), ft(0.0), ft(1.0))
-    ra_ratio = ra / max(ra + rarc, ft(1e-6))
-    
-    canopy_evap_star = (Wratio ^ (ft(2.0) / ft(3.0))) * E_p_wet * ra_ratio
-
-    # --- 5. Fraction Calculation (f_n) ---
-    # Determines what fraction of the timestep the canopy will be evaporating.
-    # If the available water (storage + new precipitation) is less than the theoretical 
-    # evaporation amount (canopy_evap_star), the canopy will dry out partway through the timestep (f_n_val < 1.0).
-    f_n_val = clamp((ws + prec) / max(canopy_evap_star, ft(1e-6)), ft(0.0), ft(1.0))
-    f_n_val = ifelse(canopy_evap_star <= ft(1e-6), ft(1.0), f_n_val)
-
-    # --- 6. Final Evaporation ---
-    # The physical evaporation that occurs during the wet fraction of the timestep.
-    evap = f_n_val * canopy_evap_star
-    
-    # Sanitize outputs to prevent NaN propagation
-    evap = ifelse(isnan(evap) | (abs(evap) > ft(1e15)), ft(0.0), evap)
-
-    return evap, f_n_val
-end
-
-# Helper functions (must be global) 
-@inline pick_evap(args...) = canopy_evap_physics_kernel(args...)[1]
-@inline pick_fn(args...)   = canopy_evap_physics_kernel(args...)[2]
-
-function calculate_canopy_evaporation!(
-    canopy_evaporation, f_n, 
-    water_storage, max_water_storage, potential_evaporation,
-    aerodynamic_resistance, rarc, prec_gpu, cv_gpu, rmin_gpu, LAI_gpu,
-    tair_gpu, elev_gpu
-)
-
-    # 1. Update canopy_evaporation
-    @. canopy_evaporation = pick_evap(
-        water_storage, max_water_storage, potential_evaporation,
-        aerodynamic_resistance, rarc, 
-        prec_gpu, LAI_gpu, tair_gpu, elev_gpu, rmin_gpu
-    )
-
-    # 2. Update f_n
-    @. f_n = pick_fn(
-        water_storage, max_water_storage, potential_evaporation,
-        aerodynamic_resistance, rarc, 
-        prec_gpu, LAI_gpu, tair_gpu, elev_gpu, rmin_gpu
-    )
-
-    # 3. Post-Process: Zero out bare soil (last index)
-    last_veg = size(canopy_evaporation, 4)
-    view(canopy_evaporation, :, :, :, last_veg) .= ft(0.0)
-
-    return nothing
-end
-
-
-@kernel function transpiration_kernel!(
-    # Outputs
-    transpiration_full,
-    transpiration_layers,
-    # Inputs
-    potential_evaporation, 
-    water_storage, 
-    max_water_storage, 
-    soil_moisture_old,       
-    soil_moisture_critical,  
-    wilting_point,           
-    root_gpu, 
-    cv_gpu, 
-    f_n,
-    AreaFract,
-    tair_gpu,    # 2D: air temperature [°C] (for Tfactor)
-    vp_gpu       # 2D: vapour pressure [kPa] (for VPDfactor)
-)
-    i, j = @index(Global, NTuple)
-
-    # Boundary Check
-    if i <= size(transpiration_full, 1) && j <= size(transpiration_full, 2)
-        
-        # Constants
-        EPS  = ft(1e-9)
-        ZERO = ft(0.0)
-        ONE  = ft(1.0)
-        
-        # --- 1. SOIL STRESS (g1, g2) ---
-        # Load Layer 1
-        W1   = soil_moisture_old[i,j,1]
-        Wcr1 = soil_moisture_critical[i,j,1]
-        Wwp1 = wilting_point[i,j,1]
-        
-        # Load Layer 2
-        W2   = soil_moisture_old[i,j,2]
-        Wcr2 = soil_moisture_critical[i,j,2]
-        Wwp2 = wilting_point[i,j,2]
-
-        # g1 = clamp((W1 - Wwp1) / (Wcr1 - Wwp1 + EPS), 0, 1)
-        g1 = clamp((W1 - Wwp1) / (Wcr1 - Wwp1 + EPS), ZERO, ONE)
-        g2 = clamp((W2 - Wwp2) / (Wcr2 - Wwp2 + EPS), ZERO, ONE)
-
-        # --- 2. VEGETATION LOOP ---
-        nveg = size(root_gpu, 4)
-        nbands = size(transpiration_full, 3)
-        
-        for k in 1:nveg
-            # Load Root Fractions
-            f1 = root_gpu[i,j,1,k]
-            f2 = root_gpu[i,j,2,k]
-            
-            # Branchless root-fraction accumulation
-            W_root_sum   = ifelse(f1 > ZERO, W1, ZERO) + ifelse(f2 > ZERO, W2, ZERO)
-            Wcr_root_sum = ifelse(f1 > ZERO, Wcr1, ZERO) + ifelse(f2 > ZERO, Wcr2, ZERO)
-            
-            share_moist = (W_root_sum >= Wcr_root_sum) & (W_root_sum > ZERO)
-            
-            moist1_wet = W1 >= Wcr1
-            moist2_wet = W2 >= Wcr2
-            
-            g_sw_veg = ifelse(share_moist, ONE, clamp((f1 * g1 + f2 * g2) / (f1 + f2 + EPS), ZERO, ONE))
-
-            e1_total = ZERO
-            e2_total = ZERO
-
-            for b in 1:nbands
-                # --- Canopy Wetness / Dry Time Factor ---
-                ws   = water_storage[i,j,b,k]
-                max_ws = max_water_storage[i,j,b,k]
-                cv   = cv_gpu[i,j,1,k]
-                fn_val = f_n[i,j,b,k]
-                pe   = potential_evaporation[i,j,b,k]
-
-                term_inner = clamp((ws / max(cv, EPS)) / max(max_ws, EPS), ZERO, ONE)
-                dry_time_factor = clamp(ONE - fn_val * (term_inner ^ (ft(2.0)/ft(3.0))), ZERO, ONE)
-
-                dry_time_factor = ifelse(k == nveg, ONE, dry_time_factor)
-
-                # --- Jarvis temperature + VPD factor for transpiration ---
-                # Tfactor: reduces transpiration in cold/hot extremes (max=1 at T=25°C)
-                # VPDfactor: CANOPY_CLOSURE=13000 Pa. Using tair_grid for PE slope
-                # gives better transpiration distribution (95.3% in run49).
-                tair_c_v     = tair_gpu[i, j]
-                Tfactor_raw  = ft(0.08) * tair_c_v - ft(0.0016) * tair_c_v * tair_c_v
-                Tfact        = Tfactor_raw < ft(1.0e-10) ? ft(1.0e-10) : (Tfactor_raw > ONE ? ONE : Tfactor_raw)
-                svp_val_t    = SVP_A * exp((SVP_B * tair_c_v) / (SVP_C + tair_c_v))  # kPa
-                vpd_pa_t     = max(svp_val_t - vp_gpu[i, j], ft(0.0)) * PA_PER_KPA  # Pa
-                raw_vpd_t    = ft(1.0) - vpd_pa_t / ft(13000.0)
-                VPDfact      = raw_vpd_t < ft(0.7) ? ft(0.7) : (raw_vpd_t > ONE ? ONE : raw_vpd_t)
-                rc_factor    = Tfact * VPDfact
-
-                # --- Transpiration Calculation ---
-                # Output aggregation in io_writer sums trans_val directly (no extra Cv weight).
-                trans_val = clamp(cv * dry_time_factor * pe * g_sw_veg * rc_factor, ZERO, ft(Inf))
-                
-                # =========================================================
-                # --- Layer Apportionment (E1, E2) without branching ---
-                # =========================================================
-                # If moisture is limited in one layer, roots can pull from another.
-                
-                # Setup base root + moisture weights
-                weight1      = f1 * g1  # root fraction * moisture stress factor
-                weight2      = f2 * g2
-                total_weight = weight1 + weight2 + EPS
-
-                # Option A: Standard Weighted Apportionment
-                # Used when moisture is severely limited (share_moist is false).
-                # Demand is strictly partitioned by the available moisture weight.
-                e1_weighted_demand = trans_val * (weight1 / total_weight)
-                e2_weighted_demand = trans_val * (weight2 / total_weight)
-
-                # Option B: Shared Moisture Apportionment (Roots redistribute uptake)
-                # Used when at least one layer has sufficient moisture (above critical point).
-                
-                # 1. Base uptake demand under shared moisture:
-                # If a layer is fully "wet", it fulfills its entire root fraction demand.
-                # If a layer is "dry", it only fulfills a stress-reduced fraction.
-                e1_shared_base = ifelse(moist1_wet, trans_val * f1, trans_val * g1 * f1)
-                e2_shared_base = ifelse(moist2_wet, trans_val * f2, trans_val * g2 * f2)
-                
-                # 2. Calculate "Spare" ET demand:
-                # This is the unmet demand from dry layers. Because roots are connected,
-                # the plant will try to pull this missing water from the wet layers instead.
-                spare_demand_from_1 = ifelse(moist1_wet, ZERO, trans_val * f1 * (ONE - g1))
-                spare_demand_from_2 = ifelse(moist2_wet, ZERO, trans_val * f2 * (ONE - g2))
-                total_spare_demand  = spare_demand_from_1 + spare_demand_from_2
-                
-                # 3. Calculate wet-layer capacity to receive the spare demand
-                # We identify which layers have surplus moisture capacity to fulfill the spare demand.
-                wet_root_frac_sum    = ifelse(moist1_wet, f1, ZERO) + ifelse(moist2_wet, f2, ZERO)
-                can_distribute_spare = (total_spare_demand > ZERO) & (wet_root_frac_sum > ZERO)
-                
-                # 4. Add the redistributed spare demand back into the wet layers proportionately
-                # The spare demand is sliced up according to the root mass present in the wet layers.
-                spare_share_e1 = ifelse(moist1_wet & can_distribute_spare, total_spare_demand * (f1 / max(wet_root_frac_sum, EPS)), ZERO)
-                spare_share_e2 = ifelse(moist2_wet & can_distribute_spare, total_spare_demand * (f2 / max(wet_root_frac_sum, EPS)), ZERO)
-                
-                e1_shared_total = e1_shared_base + spare_share_e1
-                e2_shared_total = e2_shared_base + spare_share_e2
-                
-                # Finally, select between Option A and Option B seamlessly using a branchless ternary.
-                # This eliminates massive GPU thread divergence caused by unpredictable if/else block stalling.
-                e1_val = ifelse(share_moist, e1_shared_total, e1_weighted_demand)
-                e2_val = ifelse(share_moist, e2_shared_total, e2_weighted_demand)
-
-                trans_val = ifelse(k == nveg, ZERO, trans_val)
-                e1_val    = ifelse(k == nveg, ZERO, e1_val)
-                e2_val    = ifelse(k == nveg, ZERO, e2_val)
-
-                # --- WRITE OUTPUTS ---
-                transpiration_full[i,j,b,k] = trans_val
-                
-                # Accumulate layers
-                e1_total += e1_val * AreaFract[i,j,b]
-                e2_total += e2_val * AreaFract[i,j,b]
-            end
-            
-            # 2. Layer distributed transpiration
-            transpiration_layers[i,j,1,k] = e1_total
-            transpiration_layers[i,j,2,k] = e2_total
-            transpiration_layers[i,j,3,k] = ZERO
-        end
-    end
-end
-
-function calculate_transpiration!(
-    # Outputs
-    transpiration_full, 
-    transpiration_layers, 
-    # Inputs
-    potential_evaporation, 
-    water_storage, 
-    max_water_storage, 
-    soil_moisture_old, 
-    soil_moisture_critical, 
-    wilting_point, 
-    root_gpu, 
-    cv_gpu, 
-    f_n,
-    AreaFract,
-    tair_gpu,
-    vp_gpu
-)
-    # 1. Configuration
-    kernel_launcher! = transpiration_kernel!(device_backend)    
-    nx, ny = size(transpiration_full)
-
-    # 2. Launch
-    kernel_launcher!(
-        transpiration_full, 
-        transpiration_layers,
-        potential_evaporation, 
-        water_storage, 
-        max_water_storage, 
-        soil_moisture_old, 
-        soil_moisture_critical, 
-        wilting_point, 
-        root_gpu, 
-        cv_gpu, 
-        f_n,
-        AreaFract,
-        tair_gpu,
-        vp_gpu;
-        ndrange = (nx, ny)
-    )
-
-    return nothing
-end
-
-
-
-function calculate_soil_evaporation!(
-    soil_evap,
-    soil_moisture, soil_moisture_max, potential_evaporation, 
-    b_infilt_gpu, cv_gpu, coverage_gpu, residual_moisture, AreaFract_gpu
-)
-    T = eltype(soil_evap)
-    
-    # Clear the output array first (since we accumulate into it)
-    fill!(soil_evap, ft(0.0))
-
-    # --- 1. The Scalar Physics Kernel (Inner Function) ---
-    function soil_evap_kernel(sm_top, sm_max_top, resid_top, pe, b_i, cv, cov)
-        # 1. Calculate Max Infiltration
-        max_infil = (ft(1.0) + b_i) * sm_max_top
-        
-        # 2. Moisture Ratio
-        # This fraction determines how full the theoretical soil column is.
-        ratio = clamp(ft(1.0) - sm_top / sm_max_top, ft(0.0), ft(1.0))
-        
-        # 3. Handle b_i == -1.0 case without explicit branching
-        # In VIC, if the shape parameter b_i is exactly -1.0, it denotes a linear storage limit.
-        ratio_adj = b_i == -ft(1.0) ? ratio : ratio ^ (ft(1.0) / (b_i + ft(1.0)))
-        tmp = b_i == -ft(1.0) ? max_infil : max_infil * (ft(1.0) - ratio_adj)
-
-        # 4. Saturation Check
-        is_saturated = tmp >= max_infil
-        
-        # 5. ARNO Evaporation Logic
-        # We replace the recursive if/else thresholds with branchless ternary expressions to compute the beta scaler.
-        ratio_unsat = clamp(ft(1.0) - (tmp / max_infil), ft(0.0), ft(1.0))
-        
-        ratio_powered = ratio_unsat > ft(0.0) ? ratio_unsat ^ b_i : ft(0.0)
-        as_val = ft(1.0) - ratio_powered
-        ratio_beta = ratio_powered > ft(0.0) ? ratio_powered ^ (ft(1.0) / b_i) : ft(0.0)
-        
-        # 6. Series Expansion for exact curve integration
-        dummy = ft(1.0)
-        ratio_pow_term = ratio_beta
-        for k in 1:40
-            dummy += (b_i * ratio_pow_term) / (b_i + FloatType(k))
-            ratio_pow_term *= ratio_beta
-        end
-
-        beta_asp = as_val + (ft(1.0) - as_val) * (ft(1.0) - ratio_beta) * dummy
-        
-        # 7. Final Calculation 
-        # Apply the mathematically selected multiplier (beta_asp) directly 
-        # unless full saturation naturally demands maximum potential evaporation.
-        esoil = is_saturated ? pe : pe * beta_asp
-        esoil = esoil * (ft(1.0) - cov) * cv
-        
-        # 8. Cap at Available Moisture
-        avail = max(sm_top - resid_top, ft(0.0))
-        esoil = clamp(esoil, ft(0.0), avail)
-        
-        return esoil
-    end
-
-    # --- 2. Apply Logic (Accumulate over Veg Types and Bands) ---
-    N_veg = size(cv_gpu, 4)
-    N_bands = size(AreaFract_gpu, 3)
-    for i in 1:N_veg
-        for b in 1:N_bands
-            # Esoil = Esoil_pot * beta(sm) * (1-fcanopy) * Cv per tile.
-            # pe passed in is Step 2 (snow-blended) PE which captures snow energy effects.
-            @views @. soil_evap += soil_evap_kernel(
-                soil_moisture[:,:,1],           
-                soil_moisture_max[:,:,1],       
-                residual_moisture[:,:,1],       
-                potential_evaporation[:,:,b,i],
-                b_infilt_gpu,                   
-                cv_gpu[:,:,1,i] * AreaFract_gpu[:,:,b],
-                coverage_gpu[:,:,1,i]           
-            )
-        end
-    end
-    
-    return nothing
-end
-
-function update_water_canopy_storage!(
-    water_storage, throughfall,  # Targets for mutation
-    prec, cv, canopy_evap, Wm, coverage
-)
-    
-    # 1. Update Throughfall FIRST
-    # We calculate the 'excess' logic on the fly using the *current* (old) water_storage.
-    # Logic: excess = max(0, (W + P - E) - Wm)
-    # Throughfall = (excess * coverage) + (prec * (1 - coverage))
-    @. throughfall = (max(ft(0.0), water_storage + prec - canopy_evap - Wm) * coverage) + 
-                     (prec * (ft(1.0) - coverage))
-
-    # 2. Update Water Storage SECOND
-    # Now we can safely mutate water_storage.
-    # Logic: clamped new storage
-    @. water_storage = clamp(water_storage + prec - canopy_evap, ft(0.0), Wm)
-
-    return nothing
-end
-
-# Eq. (23): Total evapotranspiration
-function calculate_total_evapotranspiration!(
-    total_et,    # Mutated Output
-    canopy_evap, transp, soil_evap, cv, coverage, AreaFract
-)
-    # 1. Initialize with Soil Evaporation
-    @. total_et = soil_evap
-
-    # 2. Accumulate Vegetation Fluxes
-    # We loop over tiles to avoid allocating massive intermediate arrays.
-    for i in 1:size(canopy_evap, 4)
-        for b in 1:size(canopy_evap, 3)
-            @views @. total_et += (canopy_evap[:,:,b,i] * cv[:,:,1,i] + transp[:,:,b,i]) * coverage[:,:,1,i] * AreaFract[:,:,b]
-        end
-    end
-    
-    return nothing
-end
